refactor: frontend API into layered structure
added clear boundaries: http/contracts/routes
