+++ docs/deployment.md
@@ -48,6 +48,13 @@ Important:
 * Do not put `VITE_*` variables in Dockge `.env`
 * They will not affect already-built bundles
 
+Frontend code boundary for API calls:
+
+* Transport: `frontend/src/lib/http/*`
+* API contracts: `frontend/src/lib/api/contracts/types.ts`
+* Route wrappers: `frontend/src/lib/api/routes/*`
+* UI imports from: `frontend/src/lib/api`
+
 If you change anything in `.env.production`, you must rebuild the frontend image.
 
 ---
