+++ frontend/src/lib/api/index.ts
@@ -1,0 +1,20 @@
+export { ApiError } from "../http/errors";
+export * from "./contracts/types";
+
+export * from "./routes/session";
+export * from "./routes/mail";
+export * from "./routes/users";
+export * from "./routes/teams";
+export * from "./routes/member";
+export * from "./routes/adminMailRequests";
+export * from "./routes/notifications";
+export * from "./routes/optix";
+
+export * as SessionApi from "./routes/session";
+export * as MailApi from "./routes/mail";
+export * as UsersApi from "./routes/users";
+export * as TeamsApi from "./routes/teams";
+export * as MemberApi from "./routes/member";
+export * as AdminMailRequestsApi from "./routes/adminMailRequests";
+export * as NotificationsApi from "./routes/notifications";
+export * as OptixApi from "./routes/optix";
