+++ frontend/src/lib/api/routes/notifications.ts
@@ -1,0 +1,17 @@
+import { apiFetch } from "../../http/client";
+import type { ApiNotifyResult } from "../contracts/types";
+
+export function sendWeeklySummaryNotification(payload: {
+  userId: string;
+  weekStart: string;
+  weekEnd: string;
+}): Promise<ApiNotifyResult> {
+  return apiFetch<ApiNotifyResult>("/api/admin/notifications/summary", {
+    method: "POST",
+    body: JSON.stringify({
+      userId: payload.userId,
+      weekStart: payload.weekStart,
+      weekEnd: payload.weekEnd,
+    }),
+  });
+}
