+++ frontend/src/lib/api/routes/teams.ts
@@ -1,0 +1,6 @@
+import { apiFetch } from "../../http/client";
+import type { ApiTeam } from "../contracts/types";
+
+export function listTeams(): Promise<ApiTeam[]> {
+  return apiFetch<ApiTeam[]>("/api/teams");
+}
