+++ frontend/src/lib/api/routes/users.ts
@@ -1,0 +1,6 @@
+import { apiFetch } from "../../http/client";
+import type { ApiUser } from "../contracts/types";
+
+export function listUsers(): Promise<ApiUser[]> {
+  return apiFetch<ApiUser[]>("/api/users");
+}
