+++ frontend/src/lib/http/errors.ts
@@ -1,0 +1,8 @@
+export class ApiError extends Error {
+  status: number;
+
+  constructor(status: number, message: string) {
+    super(message);
+    this.status = status;
+  }
+}
